[ALPS02972980] hwzram: fix unintended sign extension
Fix unintended sign extension.

MTK-Commit-Id: ae9ab74c9de2daccc1985f39412eeb09454612f9

CR-Id: ALPS02972980
Feature: UFOZIP (ZRAM Accelerator)

+++ drivers/block/zram/hwzram_impl.c
@@ -1593,7 +1593,7 @@ retrydec:
 #ifdef CONFIG_MTK_ENG_BUILD
 	csize_data = hash;
 	csize_data = (csize_data << ZRAM_DCMD_CSIZE_HASH_SHIFT) |
-		(compr_size << ZRAM_DCMD_CSIZE_SIZE_SHIFT) |
+		((uint64_t)compr_size << ZRAM_DCMD_CSIZE_SIZE_SHIFT) |
 		0x1;
 #else
 	csize_data = (uint64_t)compr_size << ZRAM_DCMD_CSIZE_SIZE_SHIFT;
